Unit test paranoia.
Some of these unit tests still crash due to issue
https://github.com/opencog/cogutil/issues/247
The few times I've been able to catch it in the act, it looked like
there still was socket data in-flight, and the that logger was logging
it, even though the sockets should have been shut down by then,
because the StorageNode dtor should have callled close(). But
somehow, this hadn't happened? Because maybe the logger was running
in a different thread? I'm very confused about what the issue is,
but the voodoo of explicitly closing the socket seems to add just
enough extra timing to mostly avoid the issue. Or that is what seems
to be happening. I's rather do this voodoo, than add sleep(1)
everywhere...