Force clean shutdown of unit test.
Really truly force a clean shutdown of the unit test.
This really truly should not be needed, but when I looked
at the logs, I saw packets that were still in flgiht, even
as the sockets had shut down. This messed with the shared
lib dtors, because code was being executed that was toching
memory that had been freed long ago. Branching to insns that
ween't mpped any more. Insanity of that magnitude.  I can't
track it down, but this seems to help qiesce the system after
the unit test has successfully completed.